testing new architecture in clustering
