+++ src/yass/cluster/__init__.py
@@ -1,7 +1,6 @@
 """
 Clustering spikes
 """
-
 from yass.cluster.run import run
 
 __all__ = ['run']
