+++ src/yass/cluster/legacy/__init__.py
@@ -1,0 +1,4 @@
+from yass.cluster.legacy.neigh_channels import neigh_channels
+from yass.cluster.legacy.location import location
+
+__all__ = ['neigh_channels', 'location']
